[skip ci] Move `.venv` to project directory
The last devcontainer created an initial `.venv` in the top level `/workspaces`
folder to avoid having the `.venv` hidden when VS Code mounted in the project
files.

Now, `poetry install` runs after the container is created so that the `.venv`
stays within the project directory. The anticipated location of the `.venv` is
added to the `PATH` during the Docker build via a build argument provided in the
`devcontainer.json` file that evaluates to the project name. As a result, this
devcontainer setup is now portable between projects!