chore: Automerge dependabot PRs via official APIs
The `merge-me` action hasn't been working for me. Let's try the
suggestions from the "Automating Dependabot" docs:
https://docs.github.com/en/code-security/dependabot/working-with-dependabot/automating-dependabot-with-github-actions#enable-auto-merge-on-a-pull-request

The new workflow will automatically apply an approval and mark the PR as
available for auto-merging once the status checks complete successfully.

While the example in the docs guards the auto-merge for a specific
dependency, this workflow allows any dependabot PR to be auto-merged.
Use the `dependabot.yml` to constrain which types of PRs dependabot can
open in the first place.